requirements.txt
polars>=0.19.0
nba_api>=1.4.1
python-dotenv>=1.0.0
pyarrow>=12.0.0
google-cloud-bigquery>=3.0.0
